+++ tests/integration/components/search-facet-registration-type-test.js
@@ -3,12 +3,18 @@ import hbs from 'htmlbars-inline-precompile';
 
 moduleForComponent('search-facet-registration-type', 'Integration | Component | search facet registration type', {
     integration: true,
+    beforeEach() {
+        this.inject.service('store', { as: 'store' });
+    },
 });
 
 test('it renders', function(assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.on('myAction', function(val) { ... });
-
+    this.set('registrationTypes', [
+        'AsPredicted Preregistration',
+        'Election Research Preacceptance Competition',
+    ]);
     this.set('facet', { key: 'registration_type', title: 'OSF Registration Type', component: 'search-facet-registration-type' });
     this.set('key', 'registration_type');
     const noop = () => {};
@@ -22,6 +28,7 @@ test('it renders', function(assert) {
         updateFilters=(action noop)
         activeFilters=activeFilters
         filterReplace=filterReplace
+        registrationTypes=registrationTypes
     }}`);
 
     assert.equal(this.$('ul > li')[0].innerText.trim(), 'AsPredicted Preregistration');